Change IoPort to architecture-independent
